pyproject.toml
[tool.poetry]
name = "mlflow-boilerplate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
pandas = "^2.1.4"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
scikit-learn = "^1.3.2"
mlflow = "^2.9.2"
ruamel-yaml = "^0.18.5"
python-box = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
